% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MiRKATMC.R
\name{MiRKATMC}
\alias{MiRKATMC}
\title{The Microbiome Regression-based Kernel Association Test for multi-categorical
(npminal or ordinal) data}
\usage{
MiRKATMC(formula, data, data.type, Ks, id = NULL, slope = NULL)
}
\arguments{
\item{formula}{the model formula with fixed effects only. The response should be
a factor.}

\item{data}{an optional data frame, ...}

\item{data.type}{either 'nominal' or 'ordinal'}

\item{Ks}{a kernel or a list of kernels}

\item{id}{an optional vector for subjects or clusters. Default is NULL for
independent response. It is required when slope is not null.}

\item{slope}{an optional vector for the random slope (e.g. time points).
Default si NULL.}
}
\value{
p value
}
\description{
MiRKATMC is based on the baseline-category logit model for nominal data or
 the proportional odds model for ordinal data, which is able to deal with
 independent and clustered/longitudinal data.
}
