% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MiRKATMC.R
\name{MiRKATMC}
\alias{MiRKATMC}
\title{The Microbiome Regression-based Kernel Association Test for multi-categorical
(nominal or ordinal) data}
\usage{
MiRKATMC(formula, data.type, Ks, random = NULL, data = NULL)
}
\arguments{
\item{formula}{The model formula with fixed effects only. The response should be
a factor.}

\item{data.type}{Either 'nominal' or 'ordinal'.}

\item{Ks}{A kernel matrix or a list of kernels.}

\item{random}{An optional formula for the random effect (default is NULL).
Currently, only models with random intercepts or with random intercepts and
slopes are supported.}

\item{data}{An optional data frame (default is NULL).}
}
\value{
A p value or a list of p values corresponding to multiple kernels. If
multiple kernels provided, the omnibus test to combine multiple p values
will be conducted and returns the p value as \code{omnibus.HMP}.
}
\description{
\code{MiRKATMC} is able to test for the association between multi-categorical
 outcomes (nominal and ordinal) and microbiome, for both independent and
 clustered/longitudinal data.
}
\examples{
test.data <- data.frame(outcome = as.factor(sample(4, 100, replace = TRUE)),
ID = gl(20, 5), time = rep(1:5, 20), age = rnorm(n = 100, mean = 30, sd = 5),
sex = rbinom(100, 1, 1/2))
D <- matrix(rbinom(10000, 2, 0.05), 100, 100)
K <- crossprod(D) # kernel matrix

# independent nominal data
MiRKATMC(formula = outcome ~ sex + age, data = test.data, data.type = 'nominal',
Ks = K)

# clustered nominal data
MiRKATMC(formula = outcome ~ sex + age, data = test.data, data.type = 'nominal',
Ks = K, random = ~ 1 | ID)

# longitudinal nominal data
MiRKATMC(formula = outcome ~ sex + age, data = test.data, data.type = 'nominal',
Ks = K, random = ~ 1 + time | ID)
}
\author{
Zhiwen Jiang
}
